+++ src/icons/filled/CopyFilledIcon.vue
@@ -1,8 +1,8 @@
 <template>
   <d-box
     is="svg"
-    width="18"
-    height="18"
+    :width="width"
+    :height="height"
     viewBox="0 0 18 18"
     fill="none"
     xmlns="http://www.w3.org/2000/svg"
@@ -45,11 +45,11 @@ defineProps({
   },
   width: {
     type: String,
-    default: "24",
+    default: "24px",
   },
   height: {
     type: String,
-    default: "24",
+    default: "24px",
   },
 });
 </script>